requirements.txt
scapy
pillow
numpy
pyqt5
matplotlib
fpdf
tensorflow
win10toast
tldextract
elasticsearch
psutil